% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_column_if_missing}
\alias{append_column_if_missing}
\title{Appends column to tibble if it doesn't exist}
\usage{
append_column_if_missing(.data, column, default_value = NA)
}
\arguments{
\item{.data}{Input data frame}

\item{column}{Column to test (unquoted)}

\item{default_value}{Value to put in column if it doesn't exist}
}
\value{
`.data` with the missing column added
}
\description{
Appends column to tibble if it doesn't exist
}
