% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{cran_all_pubs}
\alias{cran_all_pubs}
\title{3. cran_all_pubs}
\usage{
cran_all_pubs(authors)
}
\arguments{
\item{author}{A list of names of different authors.}
}
\value{
A combined dataframe of CRAN package downloads for all authors.
}
\description{
This function combines the cran publications for all authors.
}
\examples{
\dontrun{
authors_list <- list(
  list(first_name = "Michael", last_name = "Lydeamore"),
  list(first_name = "Rob", last_name = "Hyndman")
)
}
\dontrun{cran_all_pubs(authors_list)}
}
