% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_downloads.R
\name{find_cran_packages}
\alias{find_cran_packages}
\title{find_cran_packages}
\usage{
find_cran_packages(first_name, last_name)
}
\arguments{
\item{first_name}{A character string representing the author's first name.}

\item{last_name}{A character string representing the author's last name.}
}
\value{
A data frame returning package name, number of downloads,author names and last update date of the package.
}
\description{
This function searches for CRAN packages by a given author's first and last name.
}
\examples{
\dontrun{find_cran_packages("Michael", "Lydeamore")}

}
