% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_pub.R
\name{get_publications_from_orcid}
\alias{get_publications_from_orcid}
\title{get_publications_from_orcid}
\usage{
get_publications_from_orcid(orcid_ids)
}
\arguments{
\item{orcid_ids}{A character vector of ORCID IDs for which publication information is requested.}
}
\value{
A data frame containing all publications for the specified ORCID IDs.
}
\description{
The `get_publications_from_orcid` function retrieves publications for given ORCID IDs using the rorcid package.
It fetches all works associated with each ORCID ID and combines them into a single data frame.
}
\examples{
# Retrieve publications for specific ORCID IDs
get_publications_from_orcid(c("0000-0003-2531-9408", "000-0001-5738-1471"))

}
