% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_pub.R
\name{get_publications_from_scholar}
\alias{get_publications_from_scholar}
\title{get_publications_from_scholar}
\usage{
get_publications_from_scholar(scholar_id)
}
\arguments{
\item{scholar_id}{A character string representing the Google Scholar ID.}
}
\value{
A data frame containing all publications for the specified Google Scholar ID.
}
\description{
This function retrieves publications for a given Google Scholar ID and formats them into a structured data frame.
}
\examples{
\dontrun{get_publications_from_scholar("vamErfkAAAAJ")}
}
