% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{cran_all_pubs}
\alias{cran_all_pubs}
\title{cran_all_pubs}
\usage{
cran_all_pubs(authors)
}
\arguments{
\item{author}{A list of names of different authors.}
}
\value{
A combined dataframe of CRAN package downloads for all authors.
}
\description{
This function combines the cran publications for all authors.
}
\examples{
\dontrun{
cran_authors <- tibble::tibble(
  first_name = c("Michael", "Rob"),
  last_name = c("Lydeamore", "Hyndman")
)
}
\dontrun{cran_all_pubs(cran_authors)}

}
