% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{get_publications}
\alias{get_publications}
\title{Get Publications from Google Scholar and ORCID}
\usage{
get_publications(orcid_id, scholar_id)
}
\arguments{
\item{orcid_id}{ORCID ID}

\item{scholar_id}{Google Scholar ID}

\item{first_name}{Author's First Name}

\item{last_name}{Author's Last Name}
}
\value{
A list containing two dataframes: one for research outputs and one for software metadata
}
\description{
Get Publications from Google Scholar and ORCID
}
